Update
We are approaching a different type of dataset, while using a Residual Unet (2D) instead of the 3D version.

Results are promising, but the training needs more work. I'm afraid the overfitting problem will be consistent due to the nature of different data resources are limited. We'll try anything we can to achieve a reasonable result.

New model and training are in the cv-v3.ipynb file.