Training
Training now yields an acceptable prediction, you can see it in the file.

This new result is due to the ensamble of three similar models.

Anyway, I don't know what else to do to improve it, since the training is really slow either.

Next step is the web app implementation.

I'm also thinking that since the prediction works on smaller images that will be resized when sent to the backend, we will need another model for upscaling (ESREGAN).